perf(llm): load LangChain on the first call, not at import
modelFacingExtensions() installs llmExtension() for every host, so the
static `import { ChatOpenAI }` in llm-client.ts made the LSP, the MCP
server, the CLI and every REPL test load LangChain and the OpenAI SDK at
startup, whether or not they ever call a model. Importing
@repo/repl/repl cost 462ms instead of 204ms, all of it that one import.

The client now holds only type imports of LangChain, which erase, and
chatModel() awaits import("@langchain/openai") on the first real call;
langchainGenerate was already async, so nothing in the API changed. The
message-to-LangChain seam moves to its own module, since packages/ai
needs it eagerly and this package does not.

This is also what broke CI, somewhere unrelated: stream.test.ts mocks
agent.ts, which used to be the only path to the OpenAI SDK, so its first
test paid the new import inside its own 5s timeout.
